split init and new
